requirements.txt
# local package
-e .

# external requirements
black
flake8
pylint
hydra-core==1.2
omegaconf
isort
pre-commit
pytest
python-dotenv
tqdm
wandb

# generic requirements
matplotlib
numpy
pandas
scipy
seaborn
tikzplotlib

# project requirements
pytorch-lightning
torch
torchmetrics
torchvision
